# backend/go/socialnetapp/docs/swagger.yaml
definitions:
  dto.LoggedinUserDto:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/dto.UserDto'
    type: object
  dto.LoginDto:
    properties:
      id:
        type: integer
      password:
        type: string
    type: object
  dto.LoginUserDto:
    properties:
      id:
        type: integer
      password:
        type: string
    type: object
  dto.NewUserDto:
    properties:
      age:
        type: integer
      city:
        type: string
      info:
        type: string
      name:
        type: string
      password:
        type: string
      sex:
        type: string
      surname:
        type: string
    type: object
  dto.UpdatePostDto:
    properties:
      id:
        type: integer
      text:
        type: string
    type: object
  dto.UserDto:
    properties:
      age:
        type: integer
      city:
        type: string
      id:
        type: integer
      info:
        type: string
      name:
        type: string
      password:
        type: string
      sex:
        type: string
      surname:
        type: string
    type: object
  model.Message:
    properties:
      authorId:
        type: integer
      created:
        type: string
      message:
        type: string
      userId:
        type: integer
    type: object
  model.Post:
    properties:
      authorId:
        type: integer
      created:
        type: string
      id:
        type: integer
      message:
        type: string
    type: object
  model.User:
    properties:
      age:
        type: integer
      city:
        type: string
      id:
        type: integer
      info:
        type: string
      name:
        type: string
      sex:
        type: integer
      surname:
        type: string
    type: object
info:
  contact: {}
  description: This is a serice for managing users
  title: Social Net API
  version: "1.0"
paths:
  /dialog/{user_id}/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: User id
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Message'
            type: array
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Get dialog with user
      tags:
      - dialogs
  /dialog/{user_id}/send:
    post:
      consumes:
      - application/json
      parameters:
      - description: User id
        in: path
        name: user_id
        required: true
        type: string
      - description: Text
        in: body
        name: text
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Send message to user
      tags:
      - dialogs
  /friend/delete/{user_id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Friend id
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Delete user's friend
      tags:
      - friends
  /friend/set/{user_id}:
    put:
      consumes:
      - application/json
      parameters:
      - description: Friend id
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Add user's friend
      tags:
      - friends
  /login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login properties
        in: body
        name: properties
        required: true
        schema:
          $ref: '#/definitions/dto.LoginUserDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoggedinUserDto'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Login user
      tags:
      - users
  /post/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Text
        in: body
        name: text
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Add user's post
      tags:
      - posts
  /post/delete/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Post id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Delete user's post
      tags:
      - posts
  /post/feed:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Post'
            type: array
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Feed friends post
      tags:
      - posts
  /post/get/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Post id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Post'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Get user's post
      tags:
      - posts
  /post/update:
    put:
      consumes:
      - application/json
      parameters:
      - description: Post properties
        in: body
        name: properties
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePostDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Update user's post
      tags:
      - posts
  /user/get/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: User id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get user by id
      tags:
      - users
  /user/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: User properties
        in: body
        name: properties
        required: true
        schema:
          $ref: '#/definitions/dto.NewUserDto'
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Register new user
      tags:
      - users
  /user/search:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: first_name
        type: string
      - in: query
        name: last_name
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.User'
            type: array
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Find user
      tags:
      - users
  /users:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.User'
            type: array
        "500":
          description: Internal Server Error
      summary: Get all users
      tags:
      - users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"